+++ backend/src/test/java/com/kiwi/nodes/UserNodeStatusTestRepositoryInMemory.java
@@ -17,7 +17,7 @@ public class UserNodeStatusTestRepositoryInMemory implements UserNodeStatusRepos
     private final Map<UserNodeStatusKey, UserNodeStatusPersistence> store = new HashMap<>();
 
     @Override
-    public Optional<UserNodeStatusPersistence> findByIdUserIdAndIdNodeId(int userId, int nodeId) {
+    public Optional<UserNodeStatusPersistence> findByIdUserIdAndIdNodeId(Long userId, int nodeId) {
         return Optional.ofNullable(store.get(new UserNodeStatusKey(userId, nodeId)));
     }
